Site::Amazon: use DOM to scrape instead of regex
This removes most of the awful regular expressions
and uses a much cleaner DOM interface.

All white-space to spaces.
Books and Movies are parsed.
Games and Music are not yet.
Add ISBN if we can tell search string is one.
Bettery byline processing.
Add genre to games and music

Use Client HTTP agent when issuing requests to providers

Signed-off-by: Nathaniel Clark <Nathaniel.Clark@misrule.us>